Save scores to Redis
Using the new saveScore method in the redis client, check that the username from the token sent in the request matches the username sent in the request body. Then, save the new score if there is no current one or if it is higher than the current score for the given username.